Agrega identificación específica de zona muerta con escalones progresivos y su aplicación en el control del motor
- Se incorporó un tercer modo de identificación (IDEN_DEADZONE) orientado a detectar la zona muerta real del motor
- Se implementó una secuencia de escalones de bajo PWM, crecientes y simétricos en ambos sentidos
- Se agregó una fase de reposo a 0 V entre escalones para eliminar efectos de inercia
- Se permitió detectar asimetrías entre giro horario y antihorario
- Se amplió la lógica de identificación en runIdentification()
- Se mejoró la robustez de la identificación en régimen de bajo esfuerzo, evitando el sesgo que la linealización en altos PWM introduce a bajos niveles de tensión aplicada
- Se eliminó la dependencia implícita de la inercia acumulada en pruebas anteriores
- Se facilitaron futuras calibraciones de identificación modificando únicamente constantes de tiempo
- Correción en el tiempo mostrado en log

+++ src/logger.cpp
@@ -4,8 +4,9 @@
 
 
 static inline double iden_time_sec(void) {
-    return iden_counter * TS_SEC;
+    return (double)iden_counter * TS_SEC; // tiempo real en segundos
 }
+
 
 void handleLogger() {
     
@@ -24,7 +25,7 @@ void handleLogger() {
 
     if(g_iden_mode != IDEN_NONE){
         // En modo identificación, solo guardo tiempo, posición y duty
-        Serial.printf("%lu,%.4f,%.2f\n", 
+        Serial.printf("%.3f,%.4f,%.2f\n", 
                         iden_time_sec(), 
                         posicion_actual, 
                         duty_visual);
@@ -35,7 +36,7 @@ void handleLogger() {
             
         double error_visual = setpoint - posicion_actual; //Variable auxiliar para el error
             
-        Serial.printf("%lu,%.3f,%.3f,%.3f,%.2f\n", 
+        Serial.printf("%.3f,%.3f,%.3f,%.3f,%.2f\n", 
                         iden_time_sec(), 
                         setpoint, 
                         posicion_actual,
